backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
openai==1.66.2
requests==2.32.3
python-dotenv==1.0.1
pytesseract==0.3.8
Pillow==11.1.0
PyMuPDF==1.25.3
beautifulsoup4==4.13.3
pygame==2.6.1
